Split assets to separate module
